迭代4：优化打歌界面、新增结算界面、新增单曲

+++ code/View/accuracyBlock.h
@@ -9,13 +9,13 @@ Q_OBJECT
 
 public:
     explicit accuracyBlock(QWidget *parent = nullptr);
-    void setaccuracy(int accuracy) { this->accuracy = accuracy; }
+    void setaccuracy(double accuracy) { this->accuracy = accuracy; }
 
 protected:
     void paintEvent(QPaintEvent *event) override;
 
 private:
-    int accuracy = 0;
+    double accuracy = 0.0;
 };
 
 #endif // accuracyBlock_H
